services/auth-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
redis==5.2.1
aiokafka==0.12.0
structlog==24.4.0
python-dotenv==1.0.1
alembic==1.14.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
